chore(android): move/rename Android package
