+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="com.reactnativeapivideouploader">
-
+          package="video.api.reactnative.uploader">
 </manifest>
