+++ android/src/main/java/video/api/reactnative/uploader/UploaderPackage.kt
@@ -1,4 +1,4 @@
-package com.reactnativeapivideouploader
+package video.api.reactnative.uploader
 
 import com.facebook.react.ReactPackage
 import com.facebook.react.bridge.NativeModule
@@ -6,9 +6,9 @@ import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.uimanager.ViewManager
 
 
-class ApiVideoUploaderPackage : ReactPackage {
+class UploaderPackage : ReactPackage {
     override fun createNativeModules(reactContext: ReactApplicationContext): List<NativeModule> {
-        return listOf(ApiVideoUploaderModule(reactContext))
+        return listOf(UploaderModule(reactContext))
     }
 
     override fun createViewManagers(reactContext: ReactApplicationContext): List<ViewManager<*, *>> {
