+++ example/android/app/build.gradle
@@ -129,7 +129,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.example.reactnativeapivideouploader"
+        applicationId "video.api.reactnative.uploader.example"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
@@ -205,7 +205,7 @@ dependencies {
         implementation jscFlavor
     }
 
-    implementation project(':reactnativeapivideouploader')
+    implementation project(':uploader')
 }
 
 // Run this once to be able to run the application with BUCK
