+++ example/android/app/src/main/java/video/api/reactnative/uploader/example/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.reactnativeapivideouploader;
+package video.api.reactnative.uploader.example;
 
 import com.facebook.react.ReactActivity;
